Query available neurons from MouseLight site

Query tracings for neurons, and/or neuron soma locations, from the MouseLight Neuron Browser. Each neuron may have multiple tracgings, typically one for its axon and another for its dendrite.

mouselight_neuron_info(simplify = TRUE)

mouselight_neuron_somata()

Arguments

simplify
whether to simplify the query result to a simple data.frame with one row per tracing (See details).

Value

data.frame describing the available neurons. See details for the implication of the simplify argument.

Details

Mouselight neurons typically consist of two separate tracings, one for the axon and one for the dendrites. When simplify=TRUE (the default) the output is a flattened data frame with one row for every tracing. When simplify=FALSE there is one row per neuron, but the data.frame contains nested data frames with per tracing information.

The identifier which you will need to download a neuron with mouselight_read_neurons is stored in column

See also

Examples

# \donttest{
ndf=mouselight_neuron_info()
# How many tracings per neurons?
table(table(ndf$neuron.id))

#> 
#>   1   2 
#>  12 990 

# In what brain regions do neurons have their end points?
br_search = mouselight_neurons_by_brain_region("dendrite")

#> Error in mouselight_neurons_by_brain_region("dendrite"): could not find function "mouselight_neurons_by_brain_region"

View(br_search)

#> Error in as.data.frame(x): object 'br_search' not found

# }