Find out how many nodes a neuron has in each brain region

After reading tracgings from the MouseLight project (mouselight_read_brain), find out in which neuropils it has its nodes

mouselight_nodes_in_region(x, brain.areas = NULL, labels = NULL)

# S3 method for neuron
mouselight_nodes_in_region(x, brain.areas = NULL,
  labels = NULL)

# S3 method for neuronlist
mouselight_nodes_in_region(x, brain.areas = NULL,
  labels = NULL)

Arguments

x
a neuronlist of tracings that have been read into R using mouselight_read_neurons, with method = "native".

brain.areas
a vector of brain region acronyms. See mouselight_brain_region_info.

labels
a vector of label IDs, denoting neuron compartment e.g. 2 for axon, 3 for dendrite, 0 for intervening path, etc. See mouselight_read_neurons.

...
methods sent to mouselight_fetch_swc

Value

A data.frame

Details

Note, when using mouselight_neurons_by_brain_region, the brain region returned can belong to larger brain regions, which are agglomerations of smaller ones. See mouselight_brain_region_info.

See also

Examples

# \donttest{
ids=c("1ccbc478-88fb-4181-a4de-d0c22ed9738b",
  "3afa5e41-374d-45bc-a546-03f362d93649")
nn=mouselight_read_neurons(ids, method = "native")
nibr = mouselight_nodes_in_region(nn, labels = c(0, 2, 3))
print(nibr)

#>                                      6b A13 AAA ab ACA ACA1 ACA2/3 ACA5 ACA6a
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b  0   0   0  0   0    0      0    0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649  0   0   0  0   0    0      0    0     0
#>                                      ACA6b ACAd ACAd1 ACAd2/3 ACAd5 ACAd6a
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0    0     0       0     0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0    0     0       0     0      0
#>                                      ACAd6b ACAv ACAv1 ACAv2/3 ACAv5 ACAv6a
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0    0     0       0     0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0    0     0       0     0      0
#>                                      ACAv6b ACB aco act ACVI ACVII AD ADP AHA
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0   0   0   0    0     0  0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0   0   0   0    0     0  0   0   0
#>                                      AHN AHNa AHNc AHNd AHNp AI AId AId1 AId2/3
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0    0    0    0    0  0   0    0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0    0    0    0    0  0   0    0      0
#>                                      AId5 AId6a AId6b AIp AIp1 AIp2/3 AIp5
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0     0     0   0    0      0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0     0     0   0    0      0    0
#>                                      AIp6a AIp6b AIv AIv1 AIv2/3 AIv5 AIv6a
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0     0   0    0      0    0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0     0   0    0      0    0     0
#>                                      AIv6b alv AM AMB AMBd AMBv amc AMd AMv AN
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0   0  0   0    0    0   0   0   0  0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0   0  0   0    0    0   0   0   0  0
#>                                      ANcr1 ANcr1gr ANcr1mo ANcr1pu ANcr2
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0       0       0       0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0       0       0       0     0
#>                                      ANcr2gr ANcr2mo ANcr2pu AOB AOBgl AOBgr
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0       0       0   0     0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0       0       0   0     0     0
#>                                      AOBmi aolt AON AON1 AON2 AONd AONe AONl
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0    0   0    0    0    0    0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0    0   0    0    0    0    0    0
#>                                      AONm AONpv aot AP apd apf APN APr AQ ar
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0     0   0  0   0   0   0   0  0  0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0     0   0  0   0   0   0   0  0  0
#>                                      arb ARH ASO AT ATN AUD AUDd AUDd1 AUDd2/3
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0   0  0   0   0    0     0       0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0   0  0   0   0    0     0       0
#>                                      AUDd4 AUDd5 AUDd6a AUDd6b AUDp AUDp1
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0     0      0      0    0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0     0      0      0    0     0
#>                                      AUDp2/3 AUDp4 AUDp5 AUDp6a AUDp6b AUDpo
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0     0     0      0      0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0     0     0      0      0     0
#>                                      AUDpo1 AUDpo2/3 AUDpo4 AUDpo5 AUDpo6a
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0        0      0      0       0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0        0      0      0       0
#>                                      AUDpo6b AUDv AUDv1 AUDv2/3 AUDv4 AUDv5
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0    0     0       0     0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0    0     0       0     0     0
#>                                      AUDv6a AUDv6b AV AVP AVPV B BA BAC bct bic
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0      0  0   0    0 0  0   0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0      0  0   0    0 0  0   0   0   0
#>                                      BLA BLAa BLAp BLAv BMA BMAa BMAp brain BS
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0    0    0    0   0    0    0     0  0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0    0    0    0   0    0    0     0  0
#>                                      bsc BST BSTa BSTal BSTam BSTd BSTdm BSTfu
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0    0     0     0    0     0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0    0     0     0    0     0     0
#>                                      BSTif BSTju BSTmg BSTov BSTp BSTpr BSTrh
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0     0     0     0    0     0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0     0     0     0    0     0     0
#>                                      BSTse BSTtr BSTv c CA CA1 CA1slm CA1so
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0     0    0 0  0   0      0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0     0    0 0  0   0      0     0
#>                                      CA1sp CA1sr CA2 CA2slm CA2so CA2sp CA2sr
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0     0   0      0     0     0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0     0   0      0     0     0     0
#>                                      CA3 CA3slm CA3slu CA3so CA3sp CA3sr CB cbc
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0      0      0     0     0     0  0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0      0      0     0     0     0  0   0
#>                                      cbf CBN cbp cbt CBX CBXgr CBXmo CBXpu cc
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0   0   0   0     0     0     0  0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0   0   0   0     0     0     0  0
#>                                      ccb ccg ccr ccs cct CEA CEAc CEAl CEAm
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0   0   0   0   0    0    0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0   0   0   0   0    0    0    0
#>                                      CENT CENT2 CENT2gr CENT2mo CENT2pu CENT3
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0     0       0       0       0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0     0       0       0       0     0
#>                                      CENT3gr CENT3mo CENT3pu cett CH chfl chpl
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0       0       0    0  0    0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0       0       0    0  0    0    0
#>                                      cic cing CL CLA CLI cm CM CN CNlam CNspg
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0    1  0   0   0  0  0  0     0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0    1  0   0   0  0  0  0     0     0
#>                                      CNU COA COAa COAa1 COAa2 COAa3 COAp COApl
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0    0     0     0     0    0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0    0     0     0     0    0     0
#>                                      COApl1 COApl1-2 COApl1-3 COApl2 COApl3
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0        0        0      0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0        0        0      0      0
#>                                      COApm COApm1 COApm1-2 COApm1-3 COApm2
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0      0        0        0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0      0        0        0      0
#>                                      COApm3 COPY COPYgr COPYmo COPYpu CP cpd
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0    0      0      0      0  0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0    0      0      0      0  0   0
#>                                      cpt crt CS csc CSl CSm cst cstc cstu ctb
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0  0   0   0   0   0    0    0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0  0   0   0   0   0    0    0   0
#>                                      cte CTX CTXpl CTXsp CU cuf CUL CUL4
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0     0     0  0   0   0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0     0     0  0   0   0    0
#>                                      CUL4, 5 CUL4, 5gr CUL4, 5mo CUL4, 5pu
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0         0         0         0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0         0         0         0
#>                                      CUL4gr CUL4mo CUL4pu CUL5 CUL5gr CUL5mo
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0      0      0    0      0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0      0      0    0      0      0
#>                                      CUL5pu CUN cvb cVIIIn das db dc dcm DCN
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0   0   0      0   0  0  0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0   0   0      0   0  0  0   0   0
#>                                      DCO DEC DECgr DECmo DECpu df DG DG-mo
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0     0     0     0  0  0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0     0     0     0  0  0     0
#>                                      DG-po DG-sg DG-sgz DGcr DGcr-mo DGcr-po
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0     0      0    0       0       0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0     0      0    0       0       0
#>                                      DGcr-sg DGlb DGlb-mo DGlb-po DGlb-sg DGmb
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0    0       0       0       0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0    0       0       0       0    0
#>                                      DGmb-mo DGmb-po DGmb-sg dhc dl dlf DMH
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0       0       0   0  0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0       0       0   0  0   0   0
#>                                      DMHa DMHp DMHv DMX DN DORpm DORsm DP DP1
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0    0    0   0  0     0     0  0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0    0    0   0  0     0     0  0   0
#>                                      DP2 DP2/3 DP5 DP6a DR drt dscp DT dtd DTN
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0     0   0    0  0   0    0  0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0     0   0    0  0   0    0  0   0   0
#>                                      dtt ec ECO ECT ECT1 ECT2/3 ECT5 ECT6a
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0  0   0   0    0      0    0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0  0   0   0    0      0    0     0
#>                                      ECT6b ECU ee eg em ENT ENTl ENTl1 ENTl2
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0   0  0  0  0   0    0     0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0   0  0  0  0   0    0     0     0
#>                                      ENTl2/3 ENTl2a ENTl2b ENTl3 ENTl4 ENTl4/5
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0      0      0     0     0       0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0      0      0     0     0       0
#>                                      ENTl5 ENTl5/6 ENTl6a ENTl6b ENTm ENTm1
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0       0      0      0    0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0       0      0      0    0     0
#>                                      ENTm2 ENTm2a ENTm2b ENTm3 ENTm4 ENTm5
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0      0      0     0     0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0      0      0     0     0     0
#>                                      ENTm5/6 ENTm6 ENTmv ENTmv1 ENTmv2 ENTmv3
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0     0     0      0      0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0     0     0      0      0      0
#>                                      ENTmv4 ENTmv5/6 EP EPd EPI eps epsc EPv EV
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0        0  0   0   0   0    0   0  0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0        0  0   0   0   0    0   0  0
#>                                      EW fa FC FF fi fiber tracts FL FLgr FLmo
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b  0  0  0  0  0            0  0    0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649  0  0  0  0  0            0  0    0    0
#>                                      FLpu FN FOTU FOTUgr FOTUmo FOTUpu fp fpr
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0  0    0      0      0      0  0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0  0    0      0      0      0  0   0
#>                                      fr FRP FRP1 FRP2/3 FS fx fxpo fxprg fxs
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b  0   0    0      0  0  0    0     0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649  0   0    0      0  0  0    0     0   0
#>                                      GENd GENv GPe GPi GR grey grf GRN grv
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0    0   0   0  0    0   0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0    0   0   0  0    0   0   0   0
#>                                      grv of CBX grv of CTX GU GU1 GU2/3 GU4 GU5
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b          0          0  0   0     0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649          0          0  0   0     0   0   0
#>                                      GU6a GU6b gVIIn HB hbc hc HEM hf hht HIP
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0    0     0  0   0  0   0  0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0    0     0  0   0  0   0  0   0   0
#>                                      HPF HY IA IAD iaf IAM ias IB IC ICB ICc
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0  0  0   0   0   0   0  0  0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0  0  0   0   0   0   0  0  0   0   0
#>                                      ICd ICe icf icp IF IG IGL III IIIn IIn ILA
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0   0   0  0  0   0   0    0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0   0   0  0  0   0   0    0   0   0
#>                                      ILA1 ILA2 ILA2/3 ILA5 ILA6a ILA6b ILM im
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0    0      0    0     0     0   0  0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0    0      0    0     0     0   0  0
#>                                      IMD In INC int INV IO IP ipf IPF IPN IRN
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0  0   0   0   0  0  0   0   0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0  0   0   0   0  0  0   0   0   0   0
#>                                      isl islm ISN Isocortex IV IVd IVF iVIIn
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0    0   0         0  0   0   0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0    0   0         0  0   0   0     0
#>                                      IVn IXn jrb KF LA lab LAT LAV LC LD LDT
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0   0  0  0   0   0   0  0  0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0   0  0  0   0   0   0  0  0   0
#>                                      lfbs lfbst LGd LGd-co LGd-ip LGd-sh LGv
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0     0   0      0      0      0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0     0   0      0      0      0   0
#>                                      LGvl LGvm LH LHA LIN LING LINGgr LINGmo
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0    0  0   0   0    0      0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0    0  0   0   0    0      0      0
#>                                      LINGpu ll LM lot lotd lotg LP LPO LRN LRNm
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0  0  0   0    0    0  0   0   0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0  0  0   0    0    0  0   0   0    0
#>                                      LRNp LS LSc LSr LSv LSX LT LTN LZ MA MARN
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0  0   0   0   0   0  0   0  0  0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0  0   0   0   0   0  0   0  0  0    0
#>                                      MB MBmot MBO MBsen MBsta mcp mct MD MDc
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b  0     0   0     0     0   0   0  0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649  0     0   0     0     0   0   0  0   0
#>                                      MDl MDm MDRN MDRNd MDRNv ME MEA MEAad
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0    0     0     0  0   0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0    0     0     0  0   0     0
#>                                      MEAav MEApd MEApd-a MEApd-b MEApd-c MEApv
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0     0       0       0       0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0     0       0       0       0     0
#>                                      MED MEPO MEV MEZ mfb mfbc mfbs mfbse mfbsm
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0    0   0   0   0    0    0     0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0    0   0   0   0    0    0     0     0
#>                                      mfbsma mfbst mfsbshy MG MGd MGm MGv MH ml
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0     0       0  0   0   0   0  0  0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0     0       0  0   0   0   0  0  0
#>                                      mlf MM Mmme MO MO1 MO2/3 MO5 MO6a MO6b MOB
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0  0    0  0   0     0   0    0    0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0  0    0  0   0     0   0    0    0   0
#>                                      MOBgl MOBgr MOBipl MOBmi MOBopl MOp MOp1
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0     0      0     0      0   0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0     0      0     0      0   0    0
#>                                      MOp2/3 MOp5 MOp6a MOp6b MOs MOs1 MOs2/3
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0    0     0     0   0    0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0    0     0     0   0    0      0
#>                                      MOs5 MOs6a MOs6b moV mp MPN MPNc MPNl MPNm
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0     0     0   0  0   0    0    0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0     0     0   0  0   0    0    0    0
#>                                      MPO MPT MRN MRNm MRNmg MRNp MS MSC MT mtc
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0   0    0     0    0  0   0  0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0   0    0     0    0  0   0  0   0
#>                                      mtg MTN mtt mtV MV MY MY-mot MY-sat MY-sen
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0   0   0  0  0      0      0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0   0   0  0  0      0      0      0
#>                                      NB NC ND NDB nf NI NIS NLL NLLd NLLh NLLv
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b  0  0  0   0  0  0   0   0    0    0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649  0  0  0   0  0  0   0   0    0    0    0
#>                                      NLOT NLOT1 NLOT1-3 NLOT2 NLOT3 NOD NODgr
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0     0       0     0     0   0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0     0       0     0     0   0     0
#>                                      NODmo NODpu NOT NPC NR nst NTB NTS NTSce
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0     0   0   0  0   0   0   0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0     0   0   0  0   0   0   0     0
#>                                      NTSco NTSge NTSl NTSm ntt och oct OLF onl
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0     0    0    0   0   0   0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0     0    0    0   0   0   0   0   0
#>                                      OP opt or ORB ORB1 ORB2/3 ORB5 ORB6a ORB6b
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b  0   0  0   0    0      0    0     0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649  0   0  0   0    0      0    0     0     0
#>                                      ORBl ORBl1 ORBl2/3 ORBl5 ORBl6a ORBl6b
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0     0       0     0      0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0     0       0     0      0      0
#>                                      ORBm ORBm1 ORBm2 ORBm2/3 ORBm5 ORBm6a ORBv
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0     0     0       0     0      0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0     0     0       0     0      0    0
#>                                      ORBvl ORBvl1 ORBvl2/3 ORBvl5 ORBvl6a
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0      0        0      0       0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0      0        0      0       0
#>                                      ORBvl6b OT OT1 OT1-3 OT2 OT3 OV P P-mot
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0  0   0     0   0   0  0 0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0  0   0     0   0   0  0 0     0
#>                                      P-sat P-sen PA PAA PAA1 PAA1-3 PAA2 PAA3
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0     0  0   0    0      0    0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0     0  0   0    0      0    0    0
#>                                      PAG PAL PALc PALd PALm PALv pap PAR PAR1
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0    0    0    0    0   0   0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0    0    0    0    0   0   0    0
#>                                      PAR2 PAR3 PARN PAS PB PBG PBl PBlc PBld
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0    0    0   0  0   0   0    0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0    0    0   0  0   0   0    0    0
#>                                      PBle PBls PBlv PBm PBme PBmm PBmv pc pce
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0    0    0   0    0    0    0  0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0    0    0   0    0    0    0  0   0
#>                                      pcf PCG PCN PD per PERI PERI1 PERI2/3
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0   0  0   0    0     0       0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0   0  0   0    0     0       0
#>                                      PERI5 PERI6a PERI6b PF PFL PFLgr PFLmo
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0      0      0  0   0     0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0      0      0  0   0     0     0
#>                                      PFLpu pfs PG PGRN PGRNd PGRNl PH php phpd
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0   0  0    0     0     0  0   0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0   0  0    0     0     0  0   0    0
#>                                      phpl phpm phpv PHY PIN PIR PIR1 PIR1-3
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0    0    0   0   0   0    0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0    0    0   0   0   0    0      0
#>                                      PIR2 PIR3 PIS PL PL1 PL2 PL2/3 PL5 PL6a
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0    0   0  0   0   0     0   0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0    0   0  0   0   0     0   0    0
#>                                      PL6b plf pm PMd PMR pms PMv pmx PO poc POL
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0   0  0   0   0   0   0   0  0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0   0  0   0   0   0   0   0  0   0   0
#>                                      POR POST POST1 POST2 POST3 PP ppf PPN PPT
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0    0     0     0     0  0   0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0    0     0     0     0  0   0   0   0
#>                                      PPY PPYd PPYs PR PRC PRE PRE1 PRE2 PRE3
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0    0    0  0   0   0    0    0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0    0    0  0   0   0    0    0    0
#>                                      pri PRM PRMgr PRMmo PRMpu PRNc PRNr PRNv
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0     0     0     0    0    0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0     0     0     0    0    0    0
#>                                      ProS ProSd ProSd-m ProSd-sp ProSd-sr ProSv
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0     0       0        0        0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0     0       0        0        0     0
#>                                      ProSv-m ProSv-sp Prosv-sr PRP PRT PS PSCH
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0        0        0   0   0  0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0        0        0   0   0  0    0
#>                                      psf PST PSTN PSV PT ptf PTLp PTLp1 PTLp2/3
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0    0   0  0   0    0     0       0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0    0   0  0   0    0     0       0
#>                                      PTLp4 PTLp5 PTLp6a PTLp6b PVa pvbh pvbt
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0     0      0      0   0    0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0     0      0      0   0    0    0
#>                                      PVH PVHam PVHap PVHd PVHdp PVHf PVHlp PVHm
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0     0     0    0     0    0     0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0     0     0    0     0    0     0    0
#>                                      PVHmm PVHmpd PVHmpv PVHp PVHpm PVHpml
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0      0      0    0     0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0      0      0    0     0      0
#>                                      PVHpmm PVHpv PVi PVp PVpo PVR PVT PVZ py
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0     0   0   0    0   0   0   0  0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0     0   0   0    0   0   0   0  0
#>                                      pyd PYR PYRgr PYRmo PYRpu RAmb RC RCH rct
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0     0     0     0    0  0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0     0     0     0    0  0   0   0
#>                                      RE retina rf RH RHP ri RL RM RN RO RPA RPO
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b  0      0  0  0   0  0  0  0  0  0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649  0      0  0  0   0  0  0  0  0  0   0   0
#>                                      RR rrt RSP RSPagl RSPagl1 RSPagl2/3
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b  0   0   0      0       0         0
#> 3afa5e41-374d-45bc-a546-03f362d93649  0   0   0      0       0         0
#>                                      RSPagl5 RSPagl6a RSPagl6b RSPd RSPd1
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0        0        0    0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0        0        0    0     0
#>                                      RSPd2/3 RSPd4 RSPd5 RSPd6a RSPd6b RSPv
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0     0     0      0      0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0     0     0      0      0    0
#>                                      RSPv1 RSPv2 RSPv2/3 RSPv5 RSPv6a RSPv6b
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0     0       0     0      0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0     0       0     0      0      0
#>                                      rst rstl rstm RT rust SAG sAMY SBPV SCdg
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0    0    0  0    0   0    0    0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0    0    0  0    0   0    0    0    0
#>                                      SCdw SCH SCig SCig-a SCig-b SCig-c SCiw
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0   0    0      0      0      0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0   0    0      0      0      0    0
#>                                      SCm SCop scp scrt SCs SCsg sct sctd sctv
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0    0   0    0   0    0   0    0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0    0   0    0   0    0   0    0    0
#>                                      scwm SCzo sec SEZ SF SFO SG SGN SH shp SI
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0    0   0   0  0   0  0   0  0   0  0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0    0   0   0  0   0  0   0  0   0  0
#>                                      sif SIM SIMgr SIMmo SIMpu SLC SLD sm smd
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0     0     0     0   0   0  0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0     0     0     0   0   0  0   0
#>                                      SMT SNc SNl snp SNr SO SOC SOCl SOCm sop
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0   0   0   0  0   0    0    0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0   0   0   0  0   0    0    0   0
#>                                      SPA SPF SPFm SPFp SPIV sptV SPVC SPVI SPVO
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0    0    0    0    0    0    0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0    0    0    0    0    0    0    0
#>                                      SPVOcdm SPVOmdmd SPVOmdmv SPVOrdm SPVOvl
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0        0        0       0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0        0        0       0      0
#>                                      srp SS SS1 SS2/3 SS4 SS5 SS6a SS6b SSN SSp
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0  0   0     0   0   0    0    0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0  0   0     0   0   0    0    0   0   0
#>                                      SSp-bfd SSp-bfd1 SSp-bfd2/3 SSp-bfd4
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0        0          0        0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0        0          0        0
#>                                      SSp-bfd5 SSp-bfd6a SSp-bfd6b SSp-ll
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b        0         0         0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649        0         0         0      0
#>                                      SSp-ll1 SSp-ll2/3 SSp-ll4 SSp-ll5 SSp-ll6a
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0         0       0       0        0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0         0       0       0        0
#>                                      SSp-ll6b SSp-m SSp-m1 SSp-m2/3 SSp-m4
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b        0     0      0        0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649        0     0      0        0      0
#>                                      SSp-m5 SSp-m6a SSp-m6b SSp-n SSp-n1
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0       0       0     0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0       0       0     0      0
#>                                      SSp-n2/3 SSp-n4 SSp-n5 SSp-n6a SSp-n6b
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b        0      0      0       0       0
#> 3afa5e41-374d-45bc-a546-03f362d93649        0      0      0       0       0
#>                                      SSp-tr SSp-tr1 SSp-tr2/3 SSp-tr4 SSp-tr5
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0       0         0       0       0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0       0         0       0       0
#>                                      SSp-tr6a SSp-tr6b SSp-ul SSp-ul1 SSp-ul2/3
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b        0        0      0       0         0
#> 3afa5e41-374d-45bc-a546-03f362d93649        0        0      0       0         0
#>                                      SSp-ul4 SSp-ul5 SSp-ul6a SSp-ul6b SSp-un
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0       0        0        0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0       0        0        0      0
#>                                      SSp-un1 SSp-un2/3 SSp-un4 SSp-un5 SSp-un6a
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0         0       0       0        0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0         0       0       0        0
#>                                      SSp-un6b SSp1 SSp2/3 SSp4 SSp5 SSp6a SSp6b
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b        0    0      0    0    0     0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649        0    0      0    0    0     0     0
#>                                      SSs SSs1 SSs2/3 SSs4 SSs5 SSs6a SSs6b sst
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0    0      0    0    0     0     0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0    0      0    0    0     0     0   0
#>                                      st stc step stf STN stp STR STRd STRv sttl
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b  0   0    0   0   0   0   0    0    0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649  0   0    0   0   0   0   0    0    0    0
#>                                      sttv SUB SUBd SUBd-m SUBd-sp SUBd-sr SubG
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0   0    0      0       0       0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0   0    0      0       0       0    0
#>                                      SUBv SUBv-m SUBv-sp SUBv-sr SUM SUMl SUMm
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0      0       0       0   0    0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0      0       0       0   0    0    0
#>                                      sup supa supd supv SUT SUV sV svp tb tct
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0    0    0    0   0   0  0   0  0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0    0    0    0   0   0  0   0  0   0
#>                                      TEa TEa1 TEa2/3 TEa4 TEa5 TEa6a TEa6b TH
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0    0      0    0    0     0     0  0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0    0      0    0    0     0     0  0
#>                                      TM TMd TMv tn tp TR TR1 TR1-3 TR2 TR3 TRN
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b  0   0   0  0  0  0   0     0   0   0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649  0   0   0  0  0  0   0     0   0   0   0
#>                                      TRS ts tsp tspc tspd TT TTd TTd1 TTd1-4
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0  0   0    0    0  0   0    0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0  0   0    0    0  0   0    0      0
#>                                      TTd2 TTd3 TTd4 ttp TTv TTv1 TTv1-3 TTv2
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0    0    0   0   0    0      0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0    0    0   0   0    0      0    0
#>                                      TTv3 TU uf UVU UVUgr UVUmo UVUpu V V3 V4
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0  0  0   0     0     0     0 0  0  0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0  0  0   0     0     0     0 0  0  0
#>                                      V4r VAL vc VCO VENT VERM vhc VI VII VIIIn
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b   0   0  0   0    0    0   0  0   0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649   0   0  0   0    0    0   0  0   0     0
#>                                      VIIn VIn VIS VIS1 VIS2/3 VIS4 VIS5 VIS6a
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0   0   0    0      0    0    0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0   0   0    0      0    0    0     0
#>                                      VIS6b VISa VISa1 VISa2/3 VISa4 VISa5
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0    0     0       0     0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0    0     0       0     0     0
#>                                      VISa6a VISa6b VISal VISal1 VISal2/3 VISal4
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0      0     0      0        0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0      0     0      0        0      0
#>                                      VISal5 VISal6a VISal6b VISam VISam1
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0       0       0     0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0       0       0     0      0
#>                                      VISam2/3 VISam4 VISam5 VISam6a VISam6b
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b        0      0      0       0       0
#> 3afa5e41-374d-45bc-a546-03f362d93649        0      0      0       0       0
#>                                      VISC VISC1 VISC2/3 VISC4 VISC5 VISC6a
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0     0       0     0     0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0     0       0     0     0      0
#>                                      VISC6b VISl VISl1 VISl2/3 VISl4 VISl5
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0    0     0       0     0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0    0     0       0     0     0
#>                                      VISl6a VISl6b VISli VISli1 VISli2/3 VISli4
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0      0     0      0        0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0      0     0      0        0      0
#>                                      VISli5 VISli6a VISli6b VISlla VISlla1
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0       0       0      0       0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0       0       0      0       0
#>                                      VISlla2/3 VISlla4 VISlla5 VISlla6a
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b         0       0       0        0
#> 3afa5e41-374d-45bc-a546-03f362d93649         0       0       0        0
#>                                      VISlla6b VISm VISm1 VISm2/3 VISm4 VISm5
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b        0    0     0       0     0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649        0    0     0       0     0     0
#>                                      VISm6a VISm6b VISmma VISmma1 VISmma2/3
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0      0      0       0         0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0      0      0       0         0
#>                                      VISmma4 VISmma5 VISmma6a VISmma6b VISmmp
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0       0        0        0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0       0        0        0      0
#>                                      VISmmp1 VISmmp2/3 VISmmp4 VISmmp5 VISmmp6a
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0         0       0       0        0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0         0       0       0        0
#>                                      VISmmp6b VISp VISp1 VISp2/3 VISp4 VISp5
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b        0    0     0       0     0     0
#> 3afa5e41-374d-45bc-a546-03f362d93649        0    0     0       0     0     0
#>                                      VISp6a VISp6b VISpl VISpl1 VISpl2/3 VISpl4
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0      0     0      0        0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0      0     0      0        0      0
#>                                      VISpl5 VISpl6a VISpl6b VISpm VISpm1
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0       0       0     0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0       0       0     0      0
#>                                      VISpm2/3 VISpm4 VISpm5 VISpm6a VISpm6b
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b        0      0      0       0       0
#> 3afa5e41-374d-45bc-a546-03f362d93649        0      0      0       0       0
#>                                      VISpor VISpor1 VISpor2/3 VISpor4 VISpor5
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0       0         0       0       0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0       0         0       0       0
#>                                      VISpor6a VISpor6b VISrl VISrl1 VISrl2/3
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b        0        0     0      0        0
#> 3afa5e41-374d-45bc-a546-03f362d93649        0        0     0      0        0
#>                                      VISrl4 VISrl5 VISrl6a VISrl6b VISrll
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b      0      0       0       0      0
#> 3afa5e41-374d-45bc-a546-03f362d93649      0      0       0       0      0
#>                                      VISrll1 VISrll2/3 VISrll4 VISrll5 VISrll6a
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b       0         0       0       0        0
#> 3afa5e41-374d-45bc-a546-03f362d93649       0         0       0       0        0
#>                                      VISrll6b VL VLPO vlt VM VMH VMHa VMHc
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b        0  0    0   0  0   0    0    0
#> 3afa5e41-374d-45bc-a546-03f362d93649        0  0    0   0  0   0    0    0
#>                                      VMHdm VMHvl Vn VNC von VP VPL VPLpc VPM
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0     0  0   0   0  0   0     0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0     0  0   0   0  0   0     0   0
#>                                      VPMpc vrt VS vsp VTA vtd VTN vVIIIn x XII
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b     0   0  0   0   0   0   0      0 0   0
#> 3afa5e41-374d-45bc-a546-03f362d93649     0   0  0   0   0   0   0      0 0   0
#>                                      XIIn XIn Xn y z ZI
#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b    0   0  0 0 0  0
#> 3afa5e41-374d-45bc-a546-03f362d93649    0   0  0 0 0  0

# Counts in a specific brain area
cing.nodes = mouselight_nodes_in_region(nn, labels = c(0, 2, 3), brain.areas = "cing")
print(cing.nodes)

#> 1ccbc478-88fb-4181-a4de-d0c22ed9738b 3afa5e41-374d-45bc-a546-03f362d93649 
#>                                    1                                    1 

# }